+++ pingone/config.go
@@ -1,6 +1,7 @@
 package pingone
 
 import (
+	"github.com/patrickcping/pingone-go-sdk-v2/authorize"
 	"github.com/patrickcping/pingone-go-sdk-v2/management"
 	"github.com/patrickcping/pingone-go-sdk-v2/mfa"
 	"github.com/patrickcping/pingone-go-sdk-v2/pingone/model"
@@ -15,6 +16,7 @@ type Config struct {
 }
 
 type Client struct {
+	AuthorizeAPIClient  *authorize.APIClient
 	ManagementAPIClient *management.APIClient
 	MFAAPIClient        *mfa.APIClient
 	RiskAPIClient       *risk.APIClient